misc/strategies/TradingView-Machine-Learning-GUI-master/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hyperview"
version = "0.1.0"
description = "CLI-driven TradingView strategy backtester and hyper-optimizer"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "beautifulsoup4>=4.12",
    "numpy>=2.0",
    "optuna>=4.0",
    "pandas>=2.2",
    "requests>=2.31",
    "rich>=13.0",
    "TA-Lib>=0.6",
    "websocket-client>=1.7",
]

[project.scripts]
hyperview = "hyperview.cli:main"

[tool.hatch.build]
directory = ".build"

[tool.hatch.build.targets.wheel]
packages = ["hyperview", "strategy"]
